requirements.txt
streamlit==1.38.0
snowflake-connector-python==3.12.0
Pillow==10.4.0
cryptography==42.0.8
toml==0.10.2

